Day 10
Added a timer and level selection mode. Split the modes into two, a beginner one that focuses on Bohr model and the harder one, that focuses on the Lewis Dot Structures. Also fixed the ending screen to show questions correct and incorrect as well as a newly added high score tracker.